add docker file for missing tokens
