+++ src/constants.py
@@ -3,27 +3,6 @@ from pathlib import Path
 
 PROJECT_ROOT = Path(__file__).parent.parent
 
-ERC20_ABI = [
-    {
-        "constant": True,
-        "inputs": [],
-        "name": "decimals",
-        "outputs": [{"name": "", "type": "uint8"}],
-        "payable": False,
-        "stateMutability": "view",
-        "type": "function",
-    },
-    {
-        "constant": True,
-        "inputs": [],
-        "name": "symbol",
-        "outputs": [{"name": "", "type": "string"}],
-        "payable": False,
-        "stateMutability": "view",
-        "type": "function",
-    },
-]
-
 PUBLIC_RESOLVER_ABI = [
     {
         "constant": True,
